projects/hcp/packages/rahcp-etl/pyproject.toml
[project]
name = "rahcp-etl"
version = "0.1.0"
description = "Stateful ETL orchestration with JetStream"
requires-python = ">=3.13"
dependencies = ["rahcp-client", "nats-py>=2.9", "tenacity>=8.1"]

[tool.uv.sources]
rahcp-client = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
